Resize tabs only when mouse leaves tab bar
Enforce a consistent tab size and add a temporary ghost tab to the end
of the tab bar when a tab is closed, to prevent tabs from resizing
immediately.

This behavior, inspired by Chrome, makes closing a bunch of tabs in
quick succession reliable and predictable. Ideally we'd want the tabs
resizing to animate, but this is already a worthwhile improvement as it
stands.