Fix overflow bug in Chrome Beta
This is unbelievably silly, but apparently if you put a child item
that's very tall inside a flex child, and that child's ancestor doesn't
have a min-height, it won't scroll, regardless of `overflow` styles.
Instead, it just acts like `overflow: hidden`, which in this case means
that the project panel would just wander offscreen and become
unreachable for many items.

I love flexbox, but this kind of thing makes me see red.