Add print margin column setting
Previously the print margin column was taken from the wrap limit, or set
to 80 if the wrap limit was not set. This made it impossible to have a
print margin any place other than 80 when wrapping to the full window,
or a print margin column distinct from the wrapping column when using
both wrapping and a print margin.

This commit adds support for a new 'marginColumn' user setting that
allows the print margin to be set to a value independent of the wrap
limit. The print margin will still be displayed at the wrap limit column
if no specific print margin column is set.