Display distinct error message for missing font. (#670)
Previously when a font specified in user.json's fontFamily property was
missing, it was handled as if the font existed but was non-monospace,
and a dialog was brought up that displayed information that was
not adapted to that situation and potentially confusing.

This commit tests for the missing font case separately from the
non-monospace font case, and it causes a different and more appropriate
dialog to be displayed in that case.

This commit also tests whether the user selected font supports kerning
or ligatures, and if so displays the same error message displayed for
propotional fonts: It is rare and usually a bug when it happens (as
in https://github.com/googlei18n/noto-fonts/issues/1021), but it is
possible for a monospace font to have ligature or kerning pairs, and
if that happens this will cause issues similar to those that come from
using a proprotional font. This commit expands the font metric tests to
detect and handle that error case.